refactor: split progressive disclosure script
