+++ src/templates/plugin.html
@@ -3,6 +3,11 @@
 {% block head %}
     <meta name="status-badge-disabled" content="1">
     <script src="{{ url_for('static', filename='scripts/lightbox.js') }}" defer></script>
+    <script src="{{ url_for('static', filename='scripts/progressive_disclosure/utils.js') }}" defer></script>
+    <script src="{{ url_for('static', filename='scripts/progressive_disclosure/mode.js') }}" defer></script>
+    <script src="{{ url_for('static', filename='scripts/progressive_disclosure/validation_tooltips.js') }}" defer></script>
+    <script src="{{ url_for('static', filename='scripts/progressive_disclosure/wizard.js') }}" defer></script>
+    <script src="{{ url_for('static', filename='scripts/progressive_disclosure/live_preview.js') }}" defer></script>
     <script src="{{ url_for('static', filename='scripts/progressive_disclosure.js') }}" defer></script>
     <script src="{{ url_for('static', filename='scripts/plugin_form.js') }}" defer></script>
     <script src="{{ url_for('static', filename='scripts/plugin_schema.js') }}" defer></script>
